pyproject.toml
[tool.poetry]
name = "grape-kun-api"
version = "0.1.0"
description = ""
authors = ["Andrés Riquelme <andresfranco.rs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^21.1"
python-dotenv = "^0.19.0"
django-cors-headers = "^3.10.0"
manga-scrap = "^0.4.0"
django-rest-knox = "^4.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
